+++ libfreerdp/codec/ncrush.c
@@ -2942,7 +2942,7 @@ static int ncrush_generate_tables(NCRUSH_CONTEXT* context)
 		for (int j = 0; j < 1 << LOMBitsLUT[i]; j++)
 		{
 			size_t l = (cnt++) + 2ull;
-			context->HuffTableLOM[l] = (int)i;
+			context->HuffTableLOM[l] = i;
 		}
 	}
