+++ winpr/libwinpr/utils/test/TestArrayList.c
@@ -6,7 +6,7 @@
 int TestArrayList(int argc, char* argv[])
 {
 	int count = 0;
-	int rc = 0;
+	SSIZE_T rc = 0;
 	size_t val = 0;
 	wArrayList* arrayList = NULL;
 	const size_t elemsToInsert = 10;
@@ -65,7 +65,7 @@ int TestArrayList(int argc, char* argv[])
 		}
 	}
 
-	rc = ArrayList_IndexOf(arrayList, (void*)(size_t)elemsToInsert, -1, -1);
+	rc = ArrayList_IndexOf(arrayList, (void*)elemsToInsert, -1, -1);
 	printf("ArrayList index: %d\n", rc);
 	if (rc != -1)
 		return -1;
